GumaPhoto/requirements.txt
fastapi
uvicorn
jinja2
qdrant-client
pydantic
insightface
onnxruntime-gpu==1.18.1
python-multipart
imagehash
exifread
deepface
tf-keras
SQLAlchemy
Pillow
geopy
sentence-transformers

python-dotenv
google-genai
